+++ argopy/tests/test_data_fetchers.py
@@ -38,12 +38,15 @@ else:
 
 # List tests:
 def test_invalid_accesspoint():
+    src = list(AVAILABLE_SOURCES.keys())[0] # Use the first valid data source
     with pytest.raises(InvalidFetcherAccessPoint):
-        ArgoDataFetcher().invalid_accesspoint.to_xarray()
+        ArgoDataFetcher(src=src).invalid_accesspoint.to_xarray() # Can't get data if access point not defined first
+    with pytest.raises(InvalidFetcherAccessPoint):
+        ArgoDataFetcher(src=src).to_xarray() # Can't get data if access point not defined first
 
 def test_invalid_fetcher():
     with pytest.raises(InvalidFetcher):
-        ArgoDataFetcher().to_xarray() # Can't get data if access point not defined first
+        ArgoDataFetcher(src='invalid_fetcher').to_xarray()
 
 @unittest.skipUnless('localftp' in AVAILABLE_SOURCES, "requires localftp data fetcher")
 def test_unavailable_accesspoint():